+++ src/main/java/com/hfut/laboratory/controller/ProjectController.java
@@ -1,10 +1,13 @@
-package com.hfut.laboratory.controller.authc.perms;
+package com.hfut.laboratory.controller;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.hfut.laboratory.enums.ReturnCode;
 import com.hfut.laboratory.pojo.Project;
+import com.hfut.laboratory.pojo.RecordBusiness;
 import com.hfut.laboratory.service.ProjectService;
+import com.hfut.laboratory.service.RecordBusinessService;
 import com.hfut.laboratory.util.QueryWapperUtils;
 import com.hfut.laboratory.vo.ApiResponse;
 import com.hfut.laboratory.vo.PageResult;
@@ -16,7 +19,6 @@ import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -39,6 +41,9 @@ public class ProjectController {
     @Autowired
     private ProjectService projectService;
 
+    @Autowired
+    private RecordBusinessService recordBusinessService;
+
     @GetMapping("/list")
     @ApiOperation("获取收费项目列表")
     @ApiImplicitParams({
@@ -46,7 +51,7 @@ public class ProjectController {
             @ApiImplicitParam(name = "size",value = "需要数据的条数limit")
     })
     @Cacheable(value = "getProjectList",keyGenerator="simpleKeyGenerator")
-    public ApiResponse<PageResult<Project>> getProjectList(@RequestParam(required = false,defaultValue = "1") Integer current,
+    public ApiResponse getProjectList(@RequestParam(required = false,defaultValue = "1") Integer current,
                                                            @RequestParam(required = false,defaultValue = "20") Integer size){
         Page<Project> page=new Page<>(current,size);
         IPage<Project> projectIPage = projectService.page(page, null);
@@ -56,7 +61,7 @@ public class ProjectController {
     @GetMapping("/simple/list")
     @ApiOperation("获取没有冻结的收费项目列表id、name列表")
     @Cacheable(value = "getProjectSimpleList",keyGenerator="simpleKeyGenerator")
-    public ApiResponse<List<ProjectSimple>> getProjectSimpleList(){
+    public ApiResponse getProjectSimpleList(){
         List<ProjectSimple> res=new ArrayList<>();
         projectService.list(QueryWapperUtils.getInWapper("status",1))
                 .forEach(project -> res.add(new ProjectSimple(((Project)project).getId(),((Project)project).getName())));
@@ -66,7 +71,7 @@ public class ProjectController {
     @GetMapping("/simple/all/list")
     @ApiOperation("获取所有收费项目列表")
     @Cacheable(value = "getAllProjectSimpleList",keyGenerator="simpleKeyGenerator")
-    public ApiResponse<List<ProjectSimple>> getAllProjectSimpleList(){
+    public ApiResponse getAllProjectSimpleList(){
         List<ProjectSimple> res=new ArrayList<>();
         projectService.list(null).forEach(project -> res.add(new ProjectSimple(((Project)project).getId(),((Project)project).getName())));
         return ApiResponse.ok(res);
@@ -76,14 +81,14 @@ public class ProjectController {
     @ApiOperation("通过id获取收费项目")
     @ApiImplicitParam(name = "id",value = "收费项目的id")
     @Cacheable(value = "getProjectById",keyGenerator="simpleKeyGenerator")
-    public ApiResponse<Project> getProjectById(@PathVariable Integer id){
+    public ApiResponse getProjectById(@PathVariable Integer id){
         Project project = projectService.getById(id);
         return ApiResponse.ok(project);
     }
 
     @PostMapping("/freeze/{id}")
     @ApiOperation("冻结项目")
-    public ApiResponse<Void> freezeProject(@PathVariable Integer id){
+    public ApiResponse freezeProject(@PathVariable Integer id){
         Project project = projectService.getById(id);
         if(project==null){
             return ApiResponse.selfError(ReturnCode.PROJECT_NOT_EXITST);
@@ -96,7 +101,7 @@ public class ProjectController {
     @PostMapping("/add")
     @ApiOperation("添加收费项目（需要权限：[pro_add]）")
     @ApiImplicitParam(name = "project",value = "收费项目的json对象")
-    public ApiResponse<Void> insertProject(@RequestBody Project project){
+    public ApiResponse insertProject(@RequestBody Project project){
         if(project.getPercentage()==null || project.getPrice()==null ||project.getName()==null){
             return ApiResponse.selfError(ReturnCode.NEED_PARAM);
         }
@@ -108,7 +113,7 @@ public class ProjectController {
     @PutMapping("/edit")
     @ApiOperation("修改收费项目（需要权限：[pro_edit]）")
     @ApiImplicitParam(name = "project",value = "收费项目的json对象")
-    public ApiResponse<Void> updateProject(@RequestBody Project project){
+    public ApiResponse updateProject(@RequestBody Project project){
         if(project.getId()==null || project.getPercentage()==null || project.getPrice()==null ||project.getName()==null){
             return ApiResponse.selfError(ReturnCode.NEED_PARAM);
         }
@@ -124,8 +129,16 @@ public class ProjectController {
     @DeleteMapping("/del/{id}")
     @ApiOperation("删除收费项目（需要权限：[pro_del]）")
     @ApiImplicitParam(name = "id",value = "收费项目的id")
-    public ApiResponse<Void> deleteProject(@PathVariable Integer id){
+    public ApiResponse deleteProject(@PathVariable Integer id){
         boolean res=true;
+        QueryWrapper<RecordBusiness> queryWapper=new QueryWrapper<>();
+        queryWapper.and(wapper->wapper.in("type",2))
+                .and(wapper->wapper.in("thing_id",id));
+
+        if(recordBusinessService.list(queryWapper).size()!=0){
+            return ApiResponse.selfError(ReturnCode.DELETE_FALI_Foreign_KEY);
+        }
+
         try{
             res=projectService.removeById(id);
         }catch (Exception e){
